--- pool/src/certora_specs/mod.rs
@@ -1,4 +1,0 @@
-pub(crate) mod summaries;
-pub(crate) mod spec;
-pub(crate) mod token;
-pub(crate) mod model;
