--- pool/src/certora_specs/model.rs
@@ -1,19 +1,0 @@
-static mut CHECKED: bool = false;
-static mut SKOLEM_I: u32 = 0;
-
-pub fn get_checked() -> bool {
-    unsafe { CHECKED }
-}
-pub fn set_checked() {
-    unsafe { CHECKED = true }
-}
-
-pub fn skolem_i() -> u32 {
-    unsafe { SKOLEM_I }
-}
-pub fn init() {
-    unsafe {
-        SKOLEM_I = nondet::nondet();
-        CHECKED = nondet::nondet();
-    }
-}
